v2/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
drf-spectacular>=0.27
django-filter>=24.0
django-environ>=0.11
python-dotenv>=1.0
selenium>=4.18
webdriver-manager>=4.0
requests>=2.31
zstandard>=0.22
brotli>=1.1
urllib3>=2.0
